+++ src/app/iperf/iperf_units.h
@@ -19,8 +19,9 @@ enum {
 u64 unit_atof( const char *s );
 
 iperf_size_t unit_atoi( const char *s );
-//void unit_snprintf( char *s, int inLen, double inNum, char inFormat );
-void unit_snprintf( char *s, int inLen, float inNum, char inFormat );
+
+void unit_snprintf( char *s, int inLen, double inNum, char inFormat );
+//void unit_snprintf( char *s, int inLen, float inNum, char inFormat );
 
 #endif
